Implements sign out capability
Adds a bar button item to settings view that presents an alert to ask if user wants to sign out. If user selects sign out option and sign out is successful, authentication view is presented. Also fixes a bug where the collection view would be immediately instantiated even if user is signed out.