auth store
Adjusted the userReducer initialState a bit more to account for new users and to house the token to send to the server
Approximated the expected reducer cases in prep for backend work. Theyll need to be adjusted to account for the backend later.
Added a fetchUser function in actions to fetch user data, as its not sent back on login by the backend rest framework
Updated prop types in GoogleAuth and GoogleAuthContainer
Added some TODOs to fix the tests for App and Navi that broke when I actually started using redux in things
Updated loginUser in userActions to mirror changes to the reducer state