Article Modal styling
Styled the article modal. Gave it a nice little drop shadow. Not too big, but not so small as to feel like a crack in a piece of wood.
Applied global and site specific CSS resets, removing the one linked in public/index.js, and inserting the Meyer one as global and a sub site one for specific things I picked up during the design portion of FSW.
Tweaked the styling of Page.js a bit to make the preview cards feel less sharp.
Fixed a styling issue on CheckoutForm.js, allowing for the styles previously written by Eva to flourish.
Adjusted Navi.js a bit after the CSS resets threw things askew.